Add files via upload
GRiP stands for Graphical Randomness In Python and, as the name suggests, is a compilation of graphical (point 'n' click) programs written in python.

You can either download the source code and run that directly or use the installer but neither lack documentation (so much so, I had to make a separate program for it!)

NOTE: To run these programs, you will need the latest version of python installed on your system. Don't worry about missing modules, as GRiP will automatically install them when you first run it.

Enjoy! ;)